Move deduplicateCourses function to the utils directory
Since this is a helper function, we can move this to the utils directory.
