+++ src/client/components/pages/Faculty/FacultyScheduleTable.tsx
@@ -29,7 +29,6 @@ import {
 } from '@fortawesome/free-solid-svg-icons';
 import {
   FacultyAbsence,
-  FacultyCourse,
   FacultyResponseDTO,
 } from 'common/dto/faculty/FacultyResponse.dto';
 import {
@@ -45,6 +44,7 @@ import {
 import { CellLayout } from 'client/components/general';
 import { MetadataContext } from 'client/context/MetadataContext';
 import { absenceToVariant } from '../utils/absenceToVariant';
+import { deduplicateCourses } from '../utils/deduplicateCourses';
 
 /**
  * Describes the semester specific filter(s)
@@ -100,28 +100,6 @@ interface FacultyScheduleTableProps {
  */
 const computeEditAbsenceButtonId = (faculty: FacultyResponseDTO, term: TERM):
 string => `editAbsence${faculty.id}${term}`;
-
-/**
- * Computes a list of unique courses to prevent duplicate courses from showing
- */
-const deduplicateCourses = (courses: FacultyCourse[]): string[] => {
-  const uniqueCourses: string[] = [];
-  courses.forEach((course) => {
-    if (uniqueCourses.indexOf(course.catalogNumber) === -1) {
-      uniqueCourses.push(course.catalogNumber);
-      // Only parse the sameAs value if there is one
-      if (course.sameAs !== '') {
-        const sameAsCourses = course.sameAs.split(', ');
-        sameAsCourses.forEach((sameAsCourse) => {
-          if (uniqueCourses.indexOf(sameAsCourse) === -1) {
-            uniqueCourses.push(sameAsCourse);
-          }
-        });
-      }
-    }
-  });
-  return uniqueCourses;
-};
 
 /**
  * Component representing the Faculty Schedules for a given academic year
